fix: move IconButton customization in a dedicated file
